requirements.txt
aiofiles==25.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
async-timeout==5.0.1
attrs==25.4.0
bidict==0.23.1
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
coloredlogs==15.0.1
docutils==0.22.3
exceptiongroup==1.3.1
fastapi==0.124.2
flatbuffers==25.9.23
frozenlist==1.8.0
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
humanfriendly==10.0
idna==3.11
ifaddr==0.2.0
itsdangerous==2.2.0
Jinja2==3.1.6
lgpio==0.2.2.0
markdown2==2.5.4
MarkupSafe==3.0.3
mpmath==1.3.0
multidict==6.7.0
nicegui==3.3.1
numpy==1.26.4
onnxruntime==1.23.2
opencv-python==4.11.0.86
orjson==3.11.5
packaging==25.0
propcache==0.4.1
protobuf==6.33.2
psutil==7.1.3
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
python-dotenv==1.2.1
python-engineio==4.12.3
python-multipart==0.0.20
python-socketio==5.15.0
PyYAML==6.0.1
requests==2.32.5
rgpio==0.2.2.0
rpi-lgpio==0.6
RPi.GPIO==0.7.1
simple-websocket==1.1.0
starlette==0.50.0
sympy==1.14.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.1
uvicorn==0.38.0
uvloop==0.22.1
watchfiles==1.1.1
websockets==15.0.1
wsproto==1.3.2
yarl==1.22.0